Vessel particular validation changes in request file

+++ backend/Modules/Operations/Database/Migrations/2023_11_27_063228_create_ops_charterer_invoice_voyages_table.php
@@ -20,8 +20,8 @@ return new class extends Migration
             $table->string('cargo_quantity')->nullable();
             $table->string('initial_loading_point')->nullable();
             $table->string('final_loading_point')->nullable();
-            $table->float('rate_per_mt')->nullable();
-            $table->float('total_amount')->nullable();
+            $table->double('rate_per_mt')->nullable();
+            $table->double('total_amount')->nullable();
             $table->timestamps();
         });
     }
